Projecte casi acabat

+++ app/Http/Controllers/AlumneController.php
@@ -2,27 +2,66 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Response;
+
 use App\Models\Alumne;
 use App\Models\Practica;
 use Illuminate\Http\Request;
 use App\Models\CompostQuimics;
 use App\Models\Condicio;
 use App\Models\Mostra;
+use App\Models\Grup;
 use App\Models\Tasca;
 use App\Models\MostraCondComposts;
 use Illuminate\Support\Facades\Auth;
+use DateTime;
 
 
 class AlumneController extends Controller
 {
-    //
+    // Funció per ordenar les tasques per dia d'entrega
+    public function array_sort($array, $on, $order=SORT_ASC)
+    {
+    $new_array = array();
+    $sortable_array = array();
+
+    if (count($array) > 0) {
+        foreach ($array as $k => $v) {
+            if (is_array($v)) {
+                foreach ($v as $k2 => $v2) {
+                    if ($k2 == $on) {
+                        $sortable_array[$k] = $v2;
+                    }
+                }
+            } else {
+                $sortable_array[$k] = $v;
+            }
+        }
+
+        switch ($order) {
+            case SORT_ASC:
+                asort($sortable_array);
+            break;
+            case SORT_DESC:
+                arsort($sortable_array);
+            break;
+        }
+
+        foreach ($sortable_array as $k => $v) {
+            $new_array[$k] = $array[$k];
+        }
+    }
+
+    return $new_array;
+}
+
     public function listTasques()
     {
         $alumne_id = Auth::user()->alumne_id;
-        $alumne = ALumne::find($alumne_id);
+        $alumne = Alumne::find($alumne_id);
         $tasques = Tasca::all();
-        $practiques = Practica::all();
+        
+        $practique = Practica::all();
+        $practiques = $this->array_sort($practique, 'data_entrega', SORT_ASC);
         $tasquesAlumne = array();
         foreach ($tasques as $tasca) {
             if ($tasca->alumne_id == $alumne_id) {
@@ -38,27 +77,34 @@ class AlumneController extends Controller
                 }
             }
         }
+        $date = new DateTime('NOW');
+        $data = $date->format('Y-m-d');
         
-        return view('alumne.administrar_tasques', ['tasques' => $tasquesAlumne, 'practiques' => $practiques, 'alumne' => $alumne]);
+        return view('alumne.administrar_tasques', ['tasques' => $tasquesAlumne, 'data' => $data, 'practiques' => $practiques, 'alumne' => $alumne]);
     }
 
     public function realitzaTasca($id, Request $request)
     {
         if (isset($request->submit)) {
             $validated = $request->validate([
-                'nom_mostra' => 'required',
-            ]);
-            $validated = $request->validate([
-                'nom_col' => 'required',
-                'alçada_col' => 'required|numeric',
-                'temperatura' => 'required',
-                'eluent' => 'required',
-                'diametre_col' => 'required|numeric',
-                'tamany' => 'required|numeric',
-                'velocitat' => 'required',
-                'detector_uv' => 'required|numeric',
-            ]);
-
+                'nom_mostra' => 'required|max:30',
+                'nom_col' => 'required|max:30',
+                'alçada_col' => 'required|numeric|max:10000',
+                'temperatura' => 'required|max:10000|numeric',
+                'eluent' => 'required|max:25',
+                'diametre_col' => 'required|numeric|max:10000',
+                'tamany' => 'required|numeric|max:10000',
+                'velocitat' => 'required|max:10',
+                'detector_uv' => 'required|numeric|max:10000',
+                'doc' => 'max:10000'
+            ]); 
+            if (!is_string($request->doc) && $request->doc != null ){
+                if (strlen($request->doc->getClientOriginalName()) > 100){
+                    // Si el document té un nom superior a 100 caràcters.
+                    return redirect()->back();
+                }
+            }
+            
             $tasca = Tasca::find($id);
             $practId = $tasca->practica_id;
             $practica = Practica::find($practId);
@@ -96,12 +142,18 @@ class AlumneController extends Controller
                 $tasca->correcta = true;
                 $tasca->save();
             } else {
-                $condicioActual = Condicio::find($tasca->condicion_id);
-                if ($condicioActual){
+                if ($tasca->condicion_id != $condicioGuardar){
+                    $condicioActual = Condicio::find($tasca->condicion_id);
+                    if ($condicioActual){
                     $tasca->condicion_id = null;
                     $tasca->save();
                     $condicioActual->delete();
+                    }
+                }else{
+                    $tasca->condicion_id = null;
+                    $tasca->save();
                 }
+                
                 $condIncorrecta = Condicio::create([
                     'alçada_col' => $request->alçada_col,
                     "temperatura" => $request->temperatura,
@@ -117,6 +169,28 @@ class AlumneController extends Controller
                 $tasca->correcta = false;
                 $tasca->save();
             }
+            $tasca->comentari = $request->comentari;
+            $date = new DateTime('NOW');
+            $data = $date->format('Y-m-d H:i:s');
+            $tasca->data_lliurament = $data;
+
+            //Guardem el arxiu a 'documents' i li posem el nom de doc_<id_alumne>_<data_hora_segons>
+            if ($request->doc){
+                if (!is_string($request->doc) && $request->doc != null){
+                    $nomFile = $request->doc->getClientOriginalName();
+                $dataFile = explode(' ', $data);
+                $name = Auth::user()->name;
+                if (!$tasca->alumne_id){
+                    $name = Grup::find($tasca->grup_id)->nom;
+                }
+                $url = $request->file('doc')->storeAs('public', $name.'_'.$dataFile[0].'_'.$dataFile[1].'_'.$nomFile);
+                $urlValida = preg_replace('/public/', 'storage', $url);
+                $tasca->document = $urlValida;
+                }else{
+                    $tasca->document = $request->doc;
+                }
+            }
+            $tasca->save();
             return redirect()->route('tasques_alumne');
         } else {
             $tasca = Tasca::find($id);
@@ -137,8 +211,15 @@ class AlumneController extends Controller
             $condicio = Condicio::find($condicioGuardar);
             $condN = $condicio->neutre;
             $compost_quimic = CompostQuimics::all();
+            $condicioAlumne = null;
+            if ($tasca->condicion_id){
+                $condicioAlumne = Condicio::find($tasca->condicion_id);
+            }
+            $date = new DateTime('NOW');
+            $data = $date->format('Y-m-d');
 
-            return view('alumne.fer_practica', ['compost_quimic' => $compost_quimic, 'tasca_id' => $tasca->id, 'condN' => $condN, 'condicio' => $condicio, 'arrayComposts' => $arrayComposts, 'mostra' => $mostra, 'practica' => $practica]);
+            return view('alumne.fer_practica', ['condicioAlumne' => $condicioAlumne, 'compost_quimic' => $compost_quimic, 'tasca' => $tasca, 
+            'condN' => $condN, 'condicio' => $condicio, 'arrayComposts' => $arrayComposts, 'mostra' => $mostra, 'practica' => $practica, 'data' => $data]);
         }
     }
